requirements.txt
Django>=3.2.12,<3.3.0
djangorestframework>=3.13.1,<3.14.0
django-cors-headers>=3.10.0,<3.11.0
Pillow>=8.3.2,<8.4

flake8>=4.0.1,<4.1.0
